Refactor index type from u128 to u64
The code changes involve refactoring the index type from u128 to u64 in several functions. This includes the 'sample_index' function and the 'read' method of both UnstructuredStorageReader trait and SimpleTestingStorageEmulator struct. The change also simplifies the return value of 'read' method by reducing the number of Val::new calls.